Update env: mlserver split, seaborn removal and pyproject.toml refactor
