--- requirements.txt
@@ -1,8 +1,0 @@
-tqdm
-numpy
-torch==2.4.1
-torchvision==0.19.1
-cloudpickle==3.1.1
-mlflow
-mlserver_mlflow
-seaborn
